mapwithvrp/vrp-backend/requirements.txt
# --------------------------
# Core FastAPI + Web Server
# --------------------------
fastapi==0.117.1
uvicorn[standard]==0.37.0
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.5

# --------------------------
# Data Handling
# --------------------------
pandas==2.3.2
numpy==2.3.3
pytz==2025.2
python-dateutil==2.9.0.post0

# --------------------------
# VRP Solver
# --------------------------
pyvrp==0.6.3
vrplib==1.5.1

# --------------------------
# ORM + Typing
# --------------------------
pydantic==2.11.9
pydantic_core==2.33.2
typing-inspection==0.4.1
typing_extensions==4.15.0

# --------------------------
# Utils
# --------------------------
tqdm==4.67.1
anyio==4.11.0
sniffio==1.3.1
setuptools==80.9.0
wheel==0.45.1
six==1.17.0
packaging==25.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
requests
geopy
pandas
python-dotenv
